consolidated RateCombiner enum
